workflow1_endpointbuilder/sam-app/functions/function3_processobjectsandupdaterows/requirements.txt
amazon-textract-caller==0.0.13
amazon-textract-prettyprinter==0.0.9
awslambdaric==2.0.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
jmespath==0.10.0
marshmallow==3.11.1
pdf2image==1.16.0
Pillow==8.3.2
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
tabulate==0.8.9
urllib3==1.26.6
